m_samplesBuffer int32_t

+++ src/flac_decoder/flac_decoder.h
@@ -217,7 +217,7 @@ class FlacDecoder : public Decoder {
     bool            m_f_flacNewMetadataBlockPicture = false;
     bool            m_valid = false;
     uint8_t         m_flacPageNr = 0;
-    ps_ptr<int16_t> m_samplesBuffer[2];
+    ps_ptr<int32_t> m_samplesBuffer[2];
     uint16_t        m_maxBlocksize = FLAC_MAX_BLOCKSIZE;
     int32_t         m_nBytes = 0;
